review comments fix

+++ src/main/java/dev/tpcoder/coupon/model/Coupon.java
@@ -2,8 +2,9 @@ package dev.tpcoder.coupon.model;
 
 import org.springframework.data.annotation.Id;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record Coupon(@Id Long id, String title, String description, LocalDateTime startDate, LocalDateTime endDate,
-                     Operator operator, double value) {
+                     Operator operator, BigDecimal value) {
 }
